broker/cloud_functions/filter_purity/requirements.txt
google-cloud-logging
pgb-broker-utils
